Refactor: Consolidate proxy modules into unified platform package
