+++ core/__init__.py
@@ -1,5 +1,4 @@
-from .device_manager import DeviceManager, device_manager
-from .monitoring_service import MonitoringService, monitoring_service
+from unified.services import DeviceManager, device_manager, MonitoringService, monitoring_service
 
 __all__ = [
     'DeviceManager', 'device_manager',
