+++ monitoring_module/__init__.py
@@ -1,10 +1,7 @@
-"""통합 모니터링 모듈
+"""통합 모니터링 모듈 (shim to unified)"""
 
-프록시 서버의 연결 테스트, 상태 확인, 리소스 모니터링을 통합 제공합니다.
-"""
-
-from .monitor import ProxyMonitor
-from .utils import format_bytes, parse_size, get_current_timestamp
+from unified.monitoring import ProxyMonitor
+from unified.utils import format_bytes, parse_size, get_current_timestamp
 
 __all__ = [
     'ProxyMonitor',
