+++ proxy_module/__init__.py
@@ -1,6 +1,6 @@
-"""프록시 모듈 패키지"""
+"""프록시 모듈 패키지 (shim to unified)"""
 
-from .proxy_client import ProxyClient
-from .proxy_manager import ProxyManager
+from unified.proxy_client import ProxyClient
+from unified.proxy_manager import ProxyManager, proxy_manager
 
-__all__ = ['ProxyClient', 'ProxyManager']
+__all__ = ['ProxyClient', 'ProxyManager', 'proxy_manager']
