+++ requirements.txt
@@ -14,7 +14,6 @@ click==8.1.7
 itsdangerous==2.1.2
 blinker==1.7.0
 typing_extensions==4.8.0
-greenlet==3.0.1
 paramiko==3.4.0
 pysnmp==7.1.21
 pyasn1==0.6.1
